rename
패키지의 이름을 더 하는 역할에 알맞게 수정
